[#67] refactor: 상품 추가, 상품 재고 수정 API Admin 으로 이동
