--- kyumall-client/src/main/java/com/kyumall/kyumallclient/product/product/dto/CreateProductResponse.java
@@ -1,9 +1,0 @@
-package com.kyumall.kyumallclient.product.product.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter @AllArgsConstructor
-public class CreateProductResponse {
-  private Long id;
-}
